Macinabox
This is a container to be run on an Unraid server. It will help with the setup of a MacOS VM. To use this please install through Unraid's community applications but you can see the template here https://raw.githubusercontent.com/SpaceinvaderOne/Docker-Templates-Unraid/master/spaceinvaderone/macinabox.xml

Macinabox will download and install Big Sur, Catalina, Mojave or High Sierra. It will auto install the vm onto an Unraid server using Opencore as the bootloader. Macinabox installs a userscript to the server that when run will also fix any xml in the macOS vm after having been edited in the Unraid VM Manager.

If you have a version of macinabox installed before Dec 9 2020 then please make sure to delete the old container template and apdata before installing the new version.